refactor(bazel-module): Move files that support parsing to ./parser
