+++ lib/modules/manager/bazel-module/parser/context.ts
