+++ lib/modules/manager/bazel-module/parser/extension-tags.ts
@@ -1,6 +1,6 @@
 import { query as q } from 'good-enough-parser';
 import { regEx } from '../../../../util/regex';
-import type { Ctx } from '../context';
+import type { Ctx } from './context';
 import { kvParams } from './common';
 
 import { mavenExtensionPrefix, mavenExtensionTags } from './maven';
