+++ lib/modules/manager/bazel-module/parser/fragments.spec.ts
