+++ lib/modules/manager/bazel-module/parser/starlark.ts
